update Promise to OpenBSD 6.6 using a helper program
The helper program makes it easier to keep Promise up to date when
there’s a new OpenBSD release, and it gives us these for free:

• compatibility details in rustdoc (“OpenBSD 5.9–6.1, 6.3, 6.5–6.6”)
• variant identifiers guaranteed to be in lexicographic order
• variant identifiers capitalised in a deterministic way

The new variants are Error (error), Unveil (unveil), Video (video).

This is a breaking change, because it renames the following variants:
CPath (to Cpath), DiskLabel (to Disklabel), DPath (to Dpath), MCast
(to Mcast), RPath (to Rpath), TMPPath (to Tmppath), WPath (to Wpath).